requirements.txt
Flask==3.1.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Bootstrap==3.3.7.1
Flask-WTF==1.2.2
pandas
# Flask-Script